YAML is the new TOML
Decided to scrap TOM and the `Nett` library, in favor of YAML-parsing
via `YamlDotNet`. There are a few reasons for this:

- The maintainer of the Nett library has explicitly stated that he will
  not continue maintaining the library: https://github.com/paiden/Nett/issues/91

- While planning the support for handling blog posts, I realized that
  some kind of "front-matter"-like mechanism would be useful. A
  reasonably easy way to accomplish this would be by using YAML as the
  frontmatter, and I have already found a way to accomplish this by
  reading in the YamlDotNet repo.

Simply put, as much as I dislike the YAML format (yes, I genuinely do!
:-) the other options unfortunately seem even worse at this point. Using
something that is broadly maintained will be better both for me and
everyone else in the long run.