Enable Markdig extensions (#14)
The following extensions are enabled by this change:

- `.UseAdvancedExtensions()`
- `.UseSoftlineBreakAsHardlineBreak()`

The latter does have a tendency to break existing blog posts. I will go
through them and edit them in a separate commit to the `halleluja.nu`
repo.

Note that the `UseSoftLineBreakAsHardlineBreak` can potentially be
problematic, since this potentially deviates from what people expect
from a Markdown parser. I think we should make this an opt-out flag that
can you disable in either:

- `config.yaml`
- The YAML front-matter for an individual `.md` file

The config setting will serve as the global default, with the
possibility to override this for an individual blog post.