Make UseSoftLineBreakAsHardlineBreak configurable
I found an existing blog post
(https://www.halleluja.nu/undervisning-fran-bibeln/2002/2/1/du-ar-gjord-till-rattfardighet-i-kristus-jesus/)
where it simply would make little sense to spend the time to remove hard
line breaks just for the sake of it. It's better to tweak this in
`sitegen`, making it configurable so that individual blog posts can opt
out from this behavior.

Closes #15.